restructured backend project
