+++ backend/src/main/java/de/dontletyopudie/backend/controller/HomeController.java
@@ -1,6 +1,6 @@
-package de.dontletyopudie.backend;
+package de.dontletyopudie.backend.controller;
 
-import de.dontletyopudie.backend.account.AccountRepository;
+import de.dontletyopudie.backend.persistence.account.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
