+++ backend/src/main/java/de/dontletyopudie/backend/controller/JudgeController.java
@@ -1,0 +1,7 @@
+package de.dontletyopudie.backend.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class JudgeController {
+}
