+++ backend/src/main/java/de/dontletyopudie/backend/controller/UploadController.java
@@ -1,0 +1,10 @@
+package de.dontletyopudie.backend.controller;
+
+import de.dontletyopudie.backend.persistence.account.AccountRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UploadController {
+}
