+++ backend/src/main/java/de/dontletyopudie/backend/judgement/execute/JudgementExecuter.java
@@ -1,0 +1,4 @@
+package de.dontletyopudie.backend.judgement.execute;
+
+public class JudgementExecuter {
+}
