+++ backend/src/main/java/de/dontletyopudie/backend/judgement/receive/JudgementReceiver.java
@@ -1,0 +1,4 @@
+package de.dontletyopudie.backend.judgement.receive;
+
+public class JudgementReceiver {
+}
