+++ backend/src/main/java/de/dontletyopudie/backend/persistence/account/AccountRepository.java
@@ -1,4 +1,4 @@
-package de.dontletyopudie.backend.account;
+package de.dontletyopudie.backend.persistence.account;
 
 import org.springframework.data.jpa.repository.JpaRepository;
