+++ backend/src/main/java/de/dontletyopudie/backend/persistence/friend/Friend.java
@@ -1,0 +1,19 @@
+package de.dontletyopudie.backend.persistence.friend;
+
+import javax.persistence.*;
+
+@Entity(name = "Friend")
+@Table(name = "friend")
+public class Friend {
+
+    @Id
+    @SequenceGenerator(name = "friend_sequence", sequenceName = "friend_sequence",allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "friend_sequence")
+    @Column(name = "id", updatable = false)
+    private Long id;
+
+    // TODO
+
+    public Friend() {
+    }
+}
