+++ backend/src/main/java/de/dontletyopudie/backend/persistence/friend/FriendRepository.java
@@ -1,0 +1,6 @@
+package de.dontletyopudie.backend.persistence.friend;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FriendRepository extends JpaRepository<Friend, Long> {
+}
