+++ backend/src/main/java/de/dontletyopudie/backend/security/Security.java
@@ -1,0 +1,8 @@
+package de.dontletyopudie.backend.security;
+
+/**
+ * Hier passiert irgendwelcher Spring security stuff
+ * Hat aber noch keiner implementiert
+ */
+public class Security {
+}
