+++ backend/src/main/java/de/dontletyopudie/backend/upload/image/ImageReceiver.java
@@ -1,0 +1,4 @@
+package de.dontletyopudie.backend.upload.image;
+
+public class ImageReceiver {
+}
